Forward acceleration and turning based on it direction
Acceleration and turning works, but not as intented but it is a start

The turning is based on the centre of the cube rather than is back and ignores speed of car, would be perfect if it was a space ship however its not

Acceleration is done by pressing space however its intended to accelerate only when hitting as close to the beat as possible.

Will most likely migrate the parts of the script into an input handler, and having the beat checking be handled by another script along with its accuracy to then effect the acceleration of the car.

The actual turning will most likely have to change as well it controls like a spaceship