Abandoned the original car movement and marked old scripts as old
After doing a bit of looking around I realised that its potentially easier to make the car body manage its own rotation while the actual movement is a sphere collider that rolls around. I also realised that the acceleration part of the car should be handled by a seperate script as its not just holding forward to accelerate its click on the right beat to go faster.

With that being said I went back and commented out the bits of code from abandoned scripts and marked them as old in a comment, it also cleaned up the warnings I was getting from scripts that I was no longer using.