Welp it turns....kinda
Welp I got it to actually turn seems that it had something to do with the script that keep the car body upright. That being said I believe its still not functional, it will most likely need a heirarchy makeover.

Main problem is now the entire model rotates not just the turning making it not function properly (obvi)